% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/functional-spectral-density.R
\name{Epanechnikov_kernel}
\alias{Epanechnikov_kernel}
\title{The Epanechnikov weight function, with support in \eqn{[-1,1]}}
\usage{
Epanechnikov_kernel(x)
}
\arguments{
\item{x}{argument at which the function is evaluated}
}
\description{
The Epanechnikov weight function, with support in \eqn{[-1,1]}
}

