% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/ftsspec.R
\docType{package}
\name{ftsspec}
\alias{ftsspec}
\alias{ftsspec-package}
\title{ftsspec: collection of functions  for estimating spectral density operator of functional time series (FTS) and comparing the spectral density operator of two functional time series, in a way that allows detection of differences of the spectral density operator in frequencies and along the curve length.}
\description{
ftsspec: collection of functions  for estimating spectral density operator of functional time series (FTS) and comparing the spectral density operator of two functional time series, in a way that allows detection of differences of the spectral density operator in frequencies and along the curve length.
}
\references{
\cite{Tavakoli, Shahin and Panaretos, Victor M. "Detecting and Localizing Differences in Functional Time Series Dynamics: A Case Study in Molecular Biophysics", 2014, under revision}
}

