% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/functional-spectral-density.R
\name{Get_noise_sd}
\alias{Get_noise_sd}
\title{Get the square root of the covariance matrix associated to a noise type}
\usage{
Get_noise_sd(noise.type, d.n)
}
\arguments{
\item{noise.type}{the type of noise that is driving the MA process. See Details section.}

\item{d.n}{dimension of the noise that is filtered}
}
\description{
Get the square root of the covariance matrix associated to a noise type
}

