% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/functional-spectral-density.R
\name{Marginal_basis_pval}
\alias{Marginal_basis_pval}
\title{Compute the marginal  p-values at each basis coefficients of for testing the equality of two spectral density kernels}
\usage{
Marginal_basis_pval(spec1, spec2, m, kappa.square, is.pi.multiple)
}
\arguments{
\item{spec1}{The two sample spectral densities (at the same frequency \eqn{\omega}) to be compared.}

\item{spec2}{The two sample spectral densities (at the same frequency \eqn{\omega}) to be compared.}

\item{m}{The number of Fourier frequencies over which the periodogram
operator was smoothed.}

\item{kappa.square}{the L2-norm of the weight function used to estimate the
spectral density operator}

\item{is.pi.multiple}{A logical variable, to specify if \eqn{\omega = {0, \pi}} or not.}
}
\description{
Compute the marginal  p-values at each basis coefficients of for testing the equality of two spectral density kernels
}

