% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/functional-spectral-density.R
\name{plot.SampleSpec}
\alias{plot.SampleSpec}
\title{Plotting method for object inheriting from class \code{SampleSpec}}
\usage{
\method{plot}{SampleSpec}(x, ...)
}
\arguments{
\item{x}{An object of the class \code{SampleSpec}}

\item{...}{additional parameters to be passed to plot()}
}
\description{
Plotting method for object inheriting from class \code{SampleSpec}
}

