% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/functional-spectral-density.R
\name{plot.SampleSpecDiffFreqCurvelength}
\alias{plot.SampleSpecDiffFreqCurvelength}
\title{Plotting method for class \code{SampleSpecDiffFreqCurvelength}}
\usage{
\method{plot}{SampleSpecDiffFreqCurvelength}(x, ncolumns = 3, ...)
}
\arguments{
\item{x}{Object of the class \code{SampleSpecDiffFreqCurvelength}}

\item{ncolumns}{number of columns for the plots}

\item{...}{additional parameters to be passed to \code{plot()}}
}
\description{
Plotting method for class \code{SampleSpecDiffFreqCurvelength}
}

