% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/functional-spectral-density.R
\name{plot.SpecMA}
\alias{plot.SpecMA}
\title{Plotting method for object inheriting from class \code{SpecMA}}
\usage{
\method{plot}{SpecMA}(x, ...)
}
\arguments{
\item{x}{A object of the class \code{SpecMA}}

\item{...}{additional parameters to be passed to plot()}
}
\description{
Plotting method for object inheriting from class \code{SpecMA}
}

