% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/functional-spectral-density.R
\name{print.SampleSpecDiffFreqCurvelength}
\alias{print.SampleSpecDiffFreqCurvelength}
\title{Printing method for class \code{SampleSpecDiffFreqCurvelength}}
\usage{
\method{print}{SampleSpecDiffFreqCurvelength}(x, ...)
}
\arguments{
\item{x}{Object of the class \code{SampleSpecDiffFreqCurvelength}}

\item{...}{Additional arguments for print}
}
\description{
Printing method for class \code{SampleSpecDiffFreqCurvelength}
}

