% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/functional-spectral-density.R
\name{PvalAdjust}
\alias{PvalAdjust}
\alias{PvalAdjust.SampleSpecDiffFreq}
\title{Generic function to adjust pvalues}
\usage{
PvalAdjust(sample.spec.diff, method)

\method{PvalAdjust}{SampleSpecDiffFreq}(sample.spec.diff, method)
}
\arguments{
\item{sample.spec.diff}{Object of the class
\code{SampleSpecDiffFreq}}

\item{method}{method used to adjust p-values}
}
\description{
Generic function to adjust pvalues

function to adjust pvalues for class \code{SampleSpecDiffFreq}
}
\seealso{
\code{\link{Spec_compare_localize_freq}}
}

